requirements.txt
Django>=5.2.3
pytest-django==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==23.2
PyJWT==2.7.0
requests==2.32.4
psycopg2-binary==2.9.10
django-filter~=23.2
PyJWT==2.7.0
python-telegram-bot==20.3
SQLAlchemy==2.0.19
djoser==2.3.1
dotenv==0.9.9